Add UI Framework and example project
- Updated `Stardew Mods.sln` to include "UIFrameworkExample".
- Enhanced `StardewUIAPI.cs` with menu registration and update handling.
- Introduced `Origin` property in `GridLayout.cs` for positioning.
- Improved component management in `BaseMenu.cs`.
- Initialized UI Framework API in `ModEntry.cs` with example UI setup.
- Added `manifest.json` for mod metadata.
- Created `UIFrameworkExample.csproj` for project settings.
- Defined `IStardewUIAPI` interface for menu and component management.
