+++ Stardew Mods.sln
@@ -15,6 +15,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Adjustable Fertilizer", "Ad
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UIFramework", "UIFramework\UIFramework.csproj", "{1A77A589-242F-5857-480B-7E9D9C5E064A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UIFrameworkExample", "UIFrameworkExample\UIFrameworkExample.csproj", "{33D3C5BC-9A15-17B0-65A0-24CBAF16A45D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -97,6 +99,18 @@ Global
 		{1A77A589-242F-5857-480B-7E9D9C5E064A}.Release|Any CPY.Build.0 = Release|Any CPU
 		{1A77A589-242F-5857-480B-7E9D9C5E064A}.Release|x86.ActiveCfg = Release|Any CPU
 		{1A77A589-242F-5857-480B-7E9D9C5E064A}.Release|x86.Build.0 = Release|Any CPU
+		{33D3C5BC-9A15-17B0-65A0-24CBAF16A45D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{33D3C5BC-9A15-17B0-65A0-24CBAF16A45D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{33D3C5BC-9A15-17B0-65A0-24CBAF16A45D}.Debug|Any CPY.ActiveCfg = Debug|Any CPU
+		{33D3C5BC-9A15-17B0-65A0-24CBAF16A45D}.Debug|Any CPY.Build.0 = Debug|Any CPU
+		{33D3C5BC-9A15-17B0-65A0-24CBAF16A45D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{33D3C5BC-9A15-17B0-65A0-24CBAF16A45D}.Debug|x86.Build.0 = Debug|Any CPU
+		{33D3C5BC-9A15-17B0-65A0-24CBAF16A45D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{33D3C5BC-9A15-17B0-65A0-24CBAF16A45D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{33D3C5BC-9A15-17B0-65A0-24CBAF16A45D}.Release|Any CPY.ActiveCfg = Release|Any CPU
+		{33D3C5BC-9A15-17B0-65A0-24CBAF16A45D}.Release|Any CPY.Build.0 = Release|Any CPU
+		{33D3C5BC-9A15-17B0-65A0-24CBAF16A45D}.Release|x86.ActiveCfg = Release|Any CPU
+		{33D3C5BC-9A15-17B0-65A0-24CBAF16A45D}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
